% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaRMSTD.R
\name{metaRMSTD}
\alias{metaRMSTD}
\title{Meta-analysis of RMSTD at multiple time horizons}
\usage{
metaRMSTD(trialdata, time_horizons, nboot = 500, MA_method)
}
\arguments{
\item{trialdata}{IPD trial data, see details for specifications}

\item{time_horizons}{specified vector of time horizons for the meta-analysis}

\item{nboot}{the number of bootstrap iterations, if using the MVMA with bootstrap covariance matrix; default=500}

\item{MA_method}{the desired meta-analysis method; options are: "mvma", "mvma_boot", "uni", "uni_flex"}
}
\value{
The \code{metaRMSTD} function returns a list object containing the random effect model results,
the RMSTD and SE values for each trial at each available time horizon, and the estimated within-trial covariance matrix for each RCT.
}
\description{
Perform a meta-analysis with RMSTD using individual patient data.
Methods include:
\enumerate{
\item \code{"mvma"} a multivariate meta-analysis borrowing strength across time-points with within-trial covariance matrix derived analytically
\item \code{"mvma_boot"} a multivariate meta-analysis borrowing strength across time-points with within-trial covariance matrix derived by bootstrap
\item \code{"uni"} a univariate meta-analysis for combined effect at each time-point using only available data
\item \code{"uni_flex"} a univariate meta-analysis for combined effect at each time-point using estimates based on flexible parametric models
}
}
\details{
Specify the time horizons at which to calculate the meta-analytic results.
The \code{trialdata} must be formatted as a dataframe containing the IPD for each single trial including trialID indication.
Variable names must include Time ("Time"), Event status ("Event"), and randomization group ("Arm").
}
\note{
RMSTD is estimable if time horizon > minimum of
 last observed times. We implement the method of moments estimator for MVMA and Dersimonian and Laird for univariate MA
}
\references{

}
